Rydder i kode med bedre sortering og kommentarer
