+++ src/components/_common/chip/Chip.module.scss
@@ -2,10 +2,10 @@
     background-color: var(
         --navds-semantic-color-component-background-alternate
     );
+    border-radius: var(--navds-border-radius-medium);
     color: var(--navds-semantic-color-text-muted);
     display: inline-block;
     padding: 0.125rem 0.5rem;
-    border-radius: var(--navds-border-radius-medium);
 }
 
 .selected {
