+++ src/utils/scroll-to.ts
@@ -23,3 +23,19 @@ export const smoothScrollToTarget = (targetId: string, offset = 0) => {
         targetElement.focus({ preventScroll: true });
     }
 };
+
+export const isElementVisible = (element: Element, parent: Element) => {
+    if (!element.getBoundingClientRect || !parent.getBoundingClientRect) {
+        return false;
+    }
+    const elementRect = element.getBoundingClientRect();
+    const parentRect = parent.getBoundingClientRect();
+    const isVisibleHorizontally =
+        elementRect.x >= parentRect.x &&
+        elementRect.x + elementRect.width <= parentRect.x + parentRect.width;
+    const isVisibleVertically =
+        elementRect.y > parentRect.y &&
+        elementRect.y + elementRect.height < parentRect.y + parentRect.height;
+
+    return isVisibleHorizontally && isVisibleVertically;
+};
